filereader, coordinate (x,y -> lat,lon), edge ctor

+++ src/PathFinder/PathFinder/Graph.cs
@@ -8,10 +8,11 @@ namespace PathFinder
 {
     public class Graph
     {
+        /* attributes */
         public int vertexCount { get; set; }
         public List<Vertex> vertices { get; set; }
         public List<Edge> edges { get; set; }
-        public double[,]? adjacencyMatrix { get; set; }
+        public int[,]? adjacencyMatrix { get; set; }
 
         /* ctor */
         public Graph()
@@ -21,12 +22,12 @@ namespace PathFinder
             edges = new List<Edge>();
             adjacencyMatrix = null;
         }
-        public Graph(int vertexCount, List<Vertex> vertices, List<Edge> edges, double[,] adjacencyMatrix)
+        public Graph(int vertexCount, List<Vertex> vertices, List<Edge> edges, int[,] adjacencyMatrix)
         {
-        this.vertexCount = vertexCount;
-        this.vertices = vertices;
-        this.edges = edges;
-        this.adjacencyMatrix = adjacencyMatrix;
+            this.vertexCount = vertexCount;
+            this.vertices = vertices;
+            this.edges = edges;
+            this.adjacencyMatrix = adjacencyMatrix;
         }
     }
 }
